added routing from super app to revee
