+++ src/components/REVEE/ReveeMainContainer.module.css
@@ -6,5 +6,5 @@
     width: 100vw;
     height: 100vh;
     position: relative;
-    border: 1px solid red;
+    /* border: 1px solid red; */
 }
