+++ src/components/REVEE/screens/MainScreen.jsx
@@ -1,10 +1,12 @@
-import React from 'react'
+import React, { useState } from 'react'
 import style from './mainScreen.module.css'
+import BottomAd from '../advertisement/BottomAd'
 
 function MainScreen() {
+  let [enter,setEnter]=useState(false);
   return (
-    <div className={`${style.mainContainer}`}>
-          
+    <div className={`${style.mainContainer}`} on>
+          <BottomAd />
     </div>
   )
 }
