+++ src/mainContainer.module.css
@@ -1,0 +1,11 @@
+*{
+    margin: 0;
+    padding: 0;
+    font-family: Poppins;
+}
+
+.mainContainer{
+    width: 100vw;
+    height: 100vh;
+    position: relative;
+}
